tests/run-fast-snapshots-test.html

<!-- 
Copyright 2013 Google Inc. All Rights Reserved.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

    http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
-->

<!DOCTYPE html>
<title>Test flashing boxes</title>
<link rel="author" title="Sandy Phan"m href="mailto:sandyphan@google.com")>
<meta name="flags" content="dom">

<style>
  .test {
    width: 100px;
    height: 50px;
    top: 0px;
    left: 0px;
    background-color: #AFA;
    position: relative;
  }
  .rows {
    position: relative;
  }
  .expected {
    background-color: red;
    width: 100px;
    height: 50px;
    float: left;
    opacity: 0;
    position: absolute;
    top: 0px;
  }
</style>

<p>A test that take snap shots of the animation and generate another animation with boxes that flashes up</p>

<div class="rows">
  <div id="a" class="test"></div>
</div>
<div class="rows">
  <div id="b" class="test"></div>
</div>
<div class="rows">
  <div id="c" class="test"></div>
</div>


<script src="../../web-animations-js/web-animation.js"></script>
<script src="../testharness/testharness.js"></script>
<script src='../testharness/testharnessreport.js'></script>
<script src='../extra-asserts.js'></script>
<link rel='stylesheet' href='../testharness/testharness.css'>
<link rel='stylesheet' type="text/css" href="../animation-test-style.css">

<script>

/*
 * all animation scripts start here
 *
 */

var elems = document.querySelectorAll('.test');
var anims = [];

for (var i = 0; i < elems.length; i++) {
  anims[i] = new Animation(elems[i], {left: '500px'}, 5.0);
}

//var animA = new Animation(document.querySelector("#a"), {left: "500px"}, {duration: 5.0,});
//var animB = new Animation(document.querySelector("#b"), {left: "500px"}, {duration: 5.0,});
//var animC = new Animation(document.querySelector("#c"), {left: "500px"}, {duration: 5.0,});


/* 
 * all testing scripts start here
 */

function test_automated(anim, n, interval) {
  setup({timeout: 100000});

  // declare variables
  var auto = setInterval(flashing, (anim.duration * 1000) / interval);
  var j = 0;
  var target = parseInt(anim.animationFunction.frames.frames[1].value);
  var object = document.getElementsByClassName('test')[n];
  var comp = object.currentStyle || getComputedStyle(object, null);
  var vals = [];
  
  // store expected distance location values
  for (var i = 1; i <= interval; i++) {
    vals[i-1] = (target / interval) * i;
  }
  
  // create divs at appropriate locations and flash the divs
  function flashing() {
    var left = parseInt(comp.left);
    assert_approx_equals(left, vals[j], 3, "Checks that the position are relatively equals");
    
    var _newDiv = document.createElement('div');
    _newDiv.className = 'expected';
    _newDiv.style.left = left + 'px';
    document.getElementsByClassName('rows')[n].appendChild(_newDiv);
  
    var expElems = document.querySelectorAll(".expected");
    expElems[j].style.opacity = 1;
  
    setTimeout(function() {
      expElems[j].style.opacity = 0;
      j++;
    }, 40);
  }
  
  setTimeout(function() {
    clearInterval(auto);
  }, 5100);
}

test_automated(anims[0], 0, 10);
test_automated(anims[1], 1, 5);
test_automated(anims[2], 2, 3);

</script>

<!-- <div id="log"></div> -->
